feat(settings): absorb AI config into settings routes (fixes #323)
Move AI Config from standalone /ai-config route into /settings/ai-config
settings category. Settings layout updated with conditional full-width
rendering for AI config. Deleted old route, updated tests and DECISIONS.md.
