+++ tests/e2e/smoke.spec.ts
@@ -1,7 +1,14 @@
 import { test, expect } from '@playwright/test';
 
 test.describe('Smoke — all routes load', () => {
-	const routes = ['/', '/sessions', '/settings', '/workspace-settings', '/usage', '/ai-config'];
+	const routes = [
+		'/',
+		'/sessions',
+		'/settings',
+		'/workspace-settings',
+		'/usage',
+		'/settings/ai-config',
+	];
 
 	for (const route of routes) {
 		test(`${route} loads without errors`, async ({ page }) => {
